Added implementation for Response test in data layer
